Reorganize plugins layout
