+++ src/check_tokens.rs
@@ -83,6 +83,5 @@ pub fn check_tokens(
                 ));
             }
         }
-
     }
 }
