+++ src/checks_gen.rs
@@ -991,9 +991,30 @@ impl CheckCodePrefix {
             CheckCodePrefix::RUF001 => vec![CheckCode::RUF001],
             CheckCodePrefix::RUF002 => vec![CheckCode::RUF002],
             CheckCodePrefix::RUF003 => vec![CheckCode::RUF003],
-            CheckCodePrefix::S => vec![CheckCode::S101, CheckCode::S102, CheckCode::S104, CheckCode::S105, CheckCode::S106, CheckCode::S107],
-            CheckCodePrefix::S1 => vec![CheckCode::S101, CheckCode::S102, CheckCode::S104, CheckCode::S105, CheckCode::S106, CheckCode::S107],
-            CheckCodePrefix::S10 => vec![CheckCode::S101, CheckCode::S102, CheckCode::S104, CheckCode::S105, CheckCode::S106, CheckCode::S107],
+            CheckCodePrefix::S => vec![
+                CheckCode::S101,
+                CheckCode::S102,
+                CheckCode::S104,
+                CheckCode::S105,
+                CheckCode::S106,
+                CheckCode::S107,
+            ],
+            CheckCodePrefix::S1 => vec![
+                CheckCode::S101,
+                CheckCode::S102,
+                CheckCode::S104,
+                CheckCode::S105,
+                CheckCode::S106,
+                CheckCode::S107,
+            ],
+            CheckCodePrefix::S10 => vec![
+                CheckCode::S101,
+                CheckCode::S102,
+                CheckCode::S104,
+                CheckCode::S105,
+                CheckCode::S106,
+                CheckCode::S107,
+            ],
             CheckCodePrefix::S101 => vec![CheckCode::S101],
             CheckCodePrefix::S102 => vec![CheckCode::S102],
             CheckCodePrefix::S104 => vec![CheckCode::S104],
