+++ src/flake8_bandit/mod.rs
@@ -1,1 +1,2 @@
-pub mod checks;
+mod helpers;
+pub mod plugins;
