+++ src/linter.rs
@@ -468,8 +468,8 @@ mod tests {
     #[test_case(CheckCode::N816, Path::new("N816.py"); "N816")]
     #[test_case(CheckCode::N817, Path::new("N817.py"); "N817")]
     #[test_case(CheckCode::N818, Path::new("N818.py"); "N818")]
-    #[test_case(CheckCode::S101, Path::new("S101.py"); "S101")] 
-    #[test_case(CheckCode::S102, Path::new("S102.py"); "S102")] 
+    #[test_case(CheckCode::S101, Path::new("S101.py"); "S101")]
+    #[test_case(CheckCode::S102, Path::new("S102.py"); "S102")]
     #[test_case(CheckCode::S104, Path::new("S104.py"); "S104")]
     #[test_case(CheckCode::S105, Path::new("S105.py"); "S105")]
     #[test_case(CheckCode::S106, Path::new("S106.py"); "S106")]
